+++ hyjf_am_server/hyjf_am_core/hyjf_am_config/src/main/java/com/hyjf/am/config/dao/model/auto/LinkExample.java
@@ -135,52 +135,52 @@ public class LinkExample {
             return (Criteria) this;
         }
 
-        public Criteria andIdEqualTo(Short value) {
+        public Criteria andIdEqualTo(Integer value) {
             addCriterion("id =", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdNotEqualTo(Short value) {
+        public Criteria andIdNotEqualTo(Integer value) {
             addCriterion("id <>", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdGreaterThan(Short value) {
+        public Criteria andIdGreaterThan(Integer value) {
             addCriterion("id >", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdGreaterThanOrEqualTo(Short value) {
+        public Criteria andIdGreaterThanOrEqualTo(Integer value) {
             addCriterion("id >=", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdLessThan(Short value) {
+        public Criteria andIdLessThan(Integer value) {
             addCriterion("id <", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdLessThanOrEqualTo(Short value) {
+        public Criteria andIdLessThanOrEqualTo(Integer value) {
             addCriterion("id <=", value, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdIn(List<Short> values) {
+        public Criteria andIdIn(List<Integer> values) {
             addCriterion("id in", values, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdNotIn(List<Short> values) {
+        public Criteria andIdNotIn(List<Integer> values) {
             addCriterion("id not in", values, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdBetween(Short value1, Short value2) {
+        public Criteria andIdBetween(Integer value1, Integer value2) {
             addCriterion("id between", value1, value2, "id");
             return (Criteria) this;
         }
 
-        public Criteria andIdNotBetween(Short value1, Short value2) {
+        public Criteria andIdNotBetween(Integer value1, Integer value2) {
             addCriterion("id not between", value1, value2, "id");
             return (Criteria) this;
         }
@@ -315,52 +315,52 @@ public class LinkExample {
             return (Criteria) this;
         }
 
-        public Criteria andOrderEqualTo(Short value) {
+        public Criteria andOrderEqualTo(Integer value) {
             addCriterion("`order` =", value, "order");
             return (Criteria) this;
         }
 
-        public Criteria andOrderNotEqualTo(Short value) {
+        public Criteria andOrderNotEqualTo(Integer value) {
             addCriterion("`order` <>", value, "order");
             return (Criteria) this;
         }
 
-        public Criteria andOrderGreaterThan(Short value) {
+        public Criteria andOrderGreaterThan(Integer value) {
             addCriterion("`order` >", value, "order");
             return (Criteria) this;
         }
 
-        public Criteria andOrderGreaterThanOrEqualTo(Short value) {
+        public Criteria andOrderGreaterThanOrEqualTo(Integer value) {
             addCriterion("`order` >=", value, "order");
             return (Criteria) this;
         }
 
-        public Criteria andOrderLessThan(Short value) {
+        public Criteria andOrderLessThan(Integer value) {
             addCriterion("`order` <", value, "order");
             return (Criteria) this;
         }
 
-        public Criteria andOrderLessThanOrEqualTo(Short value) {
+        public Criteria andOrderLessThanOrEqualTo(Integer value) {
             addCriterion("`order` <=", value, "order");
             return (Criteria) this;
         }
 
-        public Criteria andOrderIn(List<Short> values) {
+        public Criteria andOrderIn(List<Integer> values) {
             addCriterion("`order` in", values, "order");
             return (Criteria) this;
         }
 
-        public Criteria andOrderNotIn(List<Short> values) {
+        public Criteria andOrderNotIn(List<Integer> values) {
             addCriterion("`order` not in", values, "order");
             return (Criteria) this;
         }
 
-        public Criteria andOrderBetween(Short value1, Short value2) {
+        public Criteria andOrderBetween(Integer value1, Integer value2) {
             addCriterion("`order` between", value1, value2, "order");
             return (Criteria) this;
         }
 
-        public Criteria andOrderNotBetween(Short value1, Short value2) {
+        public Criteria andOrderNotBetween(Integer value1, Integer value2) {
             addCriterion("`order` not between", value1, value2, "order");
             return (Criteria) this;
         }
